% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CI_OLS.R
\name{Compute_RnLin}
\alias{Compute_RnLin}
\title{Function to compute R_{n, lin}(gamma) for a free variable gamma
concentr_XXtranspose corresponds to the sqrt(K_reg / (n * gamma))
in the paper.
It is possible to sharpen the concentration with additional assumptions
(bounded support, etc.), see function Compute_concentrationXXt.}
\usage{
Compute_RnLin(gamma, concentr_XXtranspose, bounds, matrix_u)
}
\description{
Function to compute R_{n, lin}(gamma) for a free variable gamma
concentr_XXtranspose corresponds to the sqrt(K_reg / (n * gamma))
in the paper.
It is possible to sharpen the concentration with additional assumptions
(bounded support, etc.), see function Compute_concentrationXXt.
}
