% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CI_OLS.R
\name{Compute_Rnvar}
\alias{Compute_Rnvar}
\title{Function to compute R_{n, var}(gamma) for a free variable gamma
(The different part follows the order of the paper.)
NB: in fact, it returns ||u||^2 * Rnvar(gamma), since Rnvar is always
used multiplied by ||u||^2 (equivalently, the part ||u||^2 could be put in
the definition of Rnvar as for Rnlin).}
\usage{
Compute_Rnvar(
  gamma,
  n,
  norms_row_X,
  residuals,
  bounds,
  gammatilde,
  X,
  inverse_XXtbar,
  matrix_u
)
}
\description{
Function to compute R_{n, var}(gamma) for a free variable gamma
(The different part follows the order of the paper.)
NB: in fact, it returns ||u||^2 * Rnvar(gamma), since Rnvar is always
used multiplied by ||u||^2 (equivalently, the part ||u||^2 could be put in
the definition of Rnvar as for Rnlin).
}
