% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CI_OLS.R
\name{Compute_xi_u_for_one_obs_i}
\alias{Compute_xi_u_for_one_obs_i}
\title{Auxiliary function used for the plug-in estimate of K_xi.
Computation of xi for a given observation i
and for each u, return a vector for the different u that are
the rows of matrix_u}
\usage{
Compute_xi_u_for_one_obs_i(
  index_obs_i,
  dataX,
  inverse_XXtbar,
  matrix_u,
  residuals
)
}
\description{
Auxiliary function used for the plug-in estimate of K_xi.
Computation of xi for a given observation i
and for each u, return a vector for the different u that are
the rows of matrix_u
}
