% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CI_OLS.R
\name{concentration_Bernstein}
\alias{concentration_Bernstein}
\title{Compute the concentration bound from Lemma C.2
(Bernstein-type lemma for concentration of random, bounded matrices)}
\usage{
concentration_Bernstein(B, C, delta, n, d)
}
\arguments{
\item{B}{\eqn{:= || \E [ (A-\E[A]) (A-\E[A]) ] ||}.}

\item{C}{almost sure upper bound on the operator 2-norm of A.}

\item{delta}{confidence level}

\item{n}{sample size}

\item{d}{dimension of A}
}
\description{
Compute the concentration bound from Lemma C.2
(Bernstein-type lemma for concentration of random, bounded matrices)
}
