% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CI_mean.R
\name{Navae_ci_mean}
\alias{Navae_ci_mean}
\title{Compute NAVAE CI for the expectation based on empirical mean estimator
and Berry-Esseen (BE) or Edgeworth Expansions (EE) bounds}
\usage{
Navae_ci_mean(
  data,
  alpha = 0.05,
  a = NULL,
  bound_K = NULL,
  param_BE_EE = list(choice = "best", setup = list(continuity = FALSE, iid = TRUE,
    no_skewness = FALSE), regularity = list(C0 = 1, p = 2), eps = 0.1)
)
}
\arguments{
\item{data}{vector of univariate observations}

\item{alpha}{1 - level of confidence of the CI.
By default, alpha is set to 0.05, yielding a 95\% CI.}

\item{a}{the free parameter \eqn{a} of the interval.
If the argument is not provided, the following default choice is made
a = 1 + \eqn{b_n} with \eqn{b_n = n^{-1/5}}.
Such a choice satisfies the conditions on the sequence of free parameters
for the CI to be asymptotically exact pointwise and uniformly over the set
of distributions with a bounded kurtosis.}

\item{bound_K}{bound on the kurtosis K_4(theta) of the distribution of the
observations that are assumed to be i.i.d.
The choice of 9 covers most "usual" distribution.
Otherwise, if the argument is not provided, the value used is the plug-in
counterpart \eqn{\widehat{K}}, that is, the empirical kurtosis of the observations.}

\item{param_BE_EE}{parameters to compute the BE or EE bound \eqn{\delta_n} used
to construct the confidence interval.
If \code{param_BE_EE} is exactly equal to \code{"BE"}, then the bound used is
the best up-to-date BE bound from Shevtsova (2013) combined with a convexity
inequality.
Otherwise, \code{param_BE_EE} is a list of four objects:
1- \code{choice}:
If equal to \code{"EE"}, the bound used is Derumigny et al. (2023)'s bound
computed using the parameters specified by the rest of \code{param_BE_EE},
namely:
2- \code{setup}: itself a logical vector of size 3,
3- \code{regularity}: itself a list of length up to 3,
4- \code{eps}: value between 0 and 1/3.
as described in the arguments of the function \code{BoundEdegworth::Bound_EE1}.
Together, they specify the bounds and assumptions used to compute the
bound \eqn{\delta_n} from Derumigny et al. (2023).
Finally, if choice is equal to \code{"best"}, the bound used is the minimum between
the previous one (with \code{choice = "EE"}) and the bound \code{"BE"}.

By default, following Remark 3.3 of the article, \code{"best"} is used and
Derumigny et al. (2023)'s bounds is computed assuming i.i.d data and no other
regularity assumptions (continuous or unskewed distribution) and the bound on
kurtosis used is the one specified in the previous the argument \code{bound_K}.}
}
\value{
this function returns a numerical vector of size 2: the first element
is the lower end of the CI, the second the upper end.
In the R regime, the CI is \code{(-Inf, Inf)}.
}
\description{
Compute NAVAE CI for the expectation based on empirical mean estimator
and Berry-Esseen (BE) or Edgeworth Expansions (EE) bounds
}
\examples{
n = 1000
data = rexp(n, 1)
Navae_ci_mean(data, bound_K = 9)
Navae_ci_mean(data) # plug-in for K

n = 50 * 10^3
data = rexp(n, 1)
Navae_ci_mean(data, bound_K = 9, alpha = 0.2, a = 1 + n^(-0/5))
Navae_ci_mean(data, bound_K = 9, alpha = 0.2)
Navae_ci_mean(data, alpha = 0.2) # plug-in for K
Navae_ci_mean(data, alpha = 0.1, param_BE_EE = list(choice = "best", setup = list(continuity = FALSE, iid = TRUE, no_skewness = FALSE), regularity = list(C0 = 1, p = 2), eps = 0.1))
Navae_ci_mean(data, alpha = 0.1, param_BE_EE = list(choice = "best", setup = list(continuity = TRUE, iid = TRUE, no_skewness = FALSE), regularity = list(kappa = 0.99), eps = 0.1))
Navae_ci_mean(data, alpha = 0.05, param_BE_EE = list(choice = "best", setup = list(continuity = FALSE, iid = TRUE, no_skewness = FALSE), regularity = list(C0 = 1, p = 2), eps = 0.1))
Navae_ci_mean(data, alpha = 0.05, param_BE_EE = list(choice = "best", setup = list(continuity = TRUE, iid = TRUE, no_skewness = FALSE), regularity = list(kappa = 0.99), eps = 0.1))

n = 100 * 10^3
data = rexp(n, 1)
Navae_ci_mean(data, bound_K = 9, alpha = 0.10)
Navae_ci_mean(data, alpha = 0.10)
Navae_ci_mean(data, bound_K = 9, alpha = 0.05)
Navae_ci_mean(data, bound_K = 9, alpha = 0.05, a = 1 + n^(-1/3))
Navae_ci_mean(data, bound_K = 9, alpha = 0.05, a = 1 + n^(-1/5))
Navae_ci_mean(data, bound_K = 9, alpha = 0.05, a = 1 + n^(-0.5/5))

TODO: possibly change the choice of a when not provided to select, if exist
the best (for the resulting length of the IC) i.e. lowest a such that
we are not in the R regime?
But in all cases, it requires quite large sample sizes.

}
