% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CI_OLS.R
\name{Navae_ci_ols}
\alias{Navae_ci_ols}
\title{Compute NAVAE CI for coefficients of a linear regression based on
the OLS estimator and Berry-Esseen (BE) or Edgeworth Expansions (EE) bounds}
\usage{
Navae_ci_ols(
  Y,
  X,
  alpha = 0.05,
  a = NULL,
  power_of_n_for_b = NULL,
  omega = NULL,
  power_of_n_for_omega = NULL,
  bounds = list(lambda_reg = NULL, K_reg = NULL, K_eps = NULL, K_xi = NULL, C = NULL, B =
    NULL),
  K_xi = NULL,
  param_BE_EE = list(choice = "best", setup = list(continuity = FALSE, iid = TRUE,
    no_skewness = FALSE), regularity = list(C0 = 1, p = 2), eps = 0.1),
  options = list(center = FALSE, bounded_case = FALSE, with_Exp_regime = FALSE),
  matrix_u = diag(NCOL(X) + 1),
  verbose = FALSE
)
}
\arguments{
\item{Y}{vector of observations of the explained variables}

\item{X}{matrix of explanatory variables.
\code{X} must not contain a constant column.}

\item{alpha}{1 - level of confidence of the CI}

\item{a}{the tuning parameter \eqn{a} of the interval}

\item{omega}{the tuning parameter \eqn{\omega} of the interval}

\item{bounds}{list of bounds for the DGP.
It can contain the following items: \itemize{
   \item \code{lambda_reg}
   \item \code{K_eps}
   \item \code{K_xi}
   \item \code{K3_xi}
   \item \code{lambda3_xi}
   \item \code{K3tilde_xi}
   \item \code{B}, \code{C} Bounds for the concentration of || Xi tilde %*% Xi tilde'||
   \item \code{K_reg} Bound on
   \eqn{ E[ || vec( \widetilde{X}\widetilde{X}'- \mathbb{I}_p ) ||^2 ] }
   Defined in Assumption 3.2 (ii).
}
The bounds that are not given are replaced by plug-ins.
For K3_xi, lambda3_xi and K3tilde_xi, the bounds are obtained
from K_xi (= K4_xi)}

\item{matrix_u}{each row of this matrix is understood as a new vector u
for which a confidence interval should be computed.
By default \code{matrix_u} is the identity matrix, corresponding
to the canonical basis of \eqn{R^p}.}
}
\value{
This function returns a data frame consisting of an observation
for each vector u given as a line of \code{matrix_u},
with the following columns: \itemize{
   \item \code{lower}: lower bound of the confidence interval
   \item \code{upper}: upper bound of the confidence interval
   \item \code{regime}: the regime used for the computation of the CI.
   Four regimes are possible: \itemize{
       \item the degenerate regimes \code{R1} and \code{R2} in which
       the confidence interval is \code{(-Inf, Inf)}.
       \item the exponential regime \code{Exp}
       \item the Edgeworth regime \code{Edg}.
   }
}
}
\description{
Compute NAVAE CI for coefficients of a linear regression based on
the OLS estimator and Berry-Esseen (BE) or Edgeworth Expansions (EE) bounds
}
\examples{
n = 20000
X1 = rnorm(n, sd = 4)
true_eps = rnorm(n)
Y = 3 + 8 * X1 + true_eps
X = cbind(X1)

myCI <- CI.OLS(Y, X, alpha = 0.05, omega = 0.2, a = 2,
  bounds = list(lambda_reg = 1, K_reg = 5, K_eps = 5, K_xi = 10),
  setup = list(continuity = FALSE, no_skewness = FALSE) )

print(myCI)

myCI <- CI.OLS(Y, X, alpha = 0.05, omega = 0.2, a = 2,
  bounds = list(lambda_reg = 1, K_reg = 5, K_eps = 5, K_xi = 20),
  setup = list(continuity = TRUE, no_skewness = TRUE) )

print(myCI)

myCI <- CI.OLS(Y, X, alpha = 0.01, omega = 0.2, a = 2,
  bounds = list(lambda_reg = 1, K_reg = 5, K_eps = 5, K_xi = 10),
  setup = list(continuity = TRUE, no_skewness = TRUE) )

print(myCI)

}
