% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CI_mean_methods.R
\name{print.NAVAE_CI_Mean}
\alias{print.NAVAE_CI_Mean}
\title{Printing a NAVAE_CI_Mean object}
\usage{
\method{print}{NAVAE_CI_Mean}(x, verbose = 1, ...)
}
\arguments{
\item{x}{the object}

\item{verbose}{if zero, only basic printing is done. Higher values corresponds
to more detailed output.}
}
\value{
\code{print.Navae_ci_ols} prints information about \code{x} and returns it
invisibly.

\code{print.Navae_ci_ols} prints information about \code{x} and returns it
invisibly.
}
\description{
Printing a NAVAE_CI_Mean object
}
