% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CI_mean_methods.R
\name{print.NAVAE_CI_Mean}
\alias{print.NAVAE_CI_Mean}
\alias{as.data.frame.NAVAE_CI_Mean}
\title{Printing a NAVAE_CI_Mean object}
\usage{
\method{print}{NAVAE_CI_Mean}(x, verbose = 0, ...)

\method{as.data.frame}{NAVAE_CI_Mean}(x, ...)
}
\arguments{
\item{x}{the object}

\item{verbose}{if zero, only basic printing is done. Higher values corresponds
to more detailed output.}

\item{...}{other arguments, currently ignored.}
}
\value{
\code{print.Navae_ci_ols} prints information about \code{x} and returns it
invisibly.

\code{as.data.frame} returns a dataframe with 2 rows.
}
\description{
Printing a NAVAE_CI_Mean object
}
\examples{
n = 10000
x = rexp(n, 1)
myCI = Navae_ci_mean(x, bound_K = 9, alpha = 0.2)

print(myCI)
as.data.frame(myCI)


}
